+++ counterweight/styles/styles.py
@@ -2,15 +2,12 @@ from __future__ import annotations
 
 from enum import Enum
 from functools import cached_property, lru_cache
-from typing import TYPE_CHECKING, Literal, NamedTuple, TypeVar
+from typing import Literal, NamedTuple, TypeVar
 
 from cachetools import LRUCache
 from pydantic import Field, NonNegativeInt, PositiveInt
 
 from counterweight.types import FrozenForbidExtras
-
-if TYPE_CHECKING:
-    pass
 
 S = TypeVar("S", bound="StyleFragment")
 
@@ -62,20 +59,37 @@ class StyleFragment(FrozenForbidExtras):
             STYLE_MERGE_CACHE[key] = merged
             return merged
 
-    def mergeable_dump(self) -> dict[str, object]:
-        d = super().model_dump(exclude_unset=True)
-
-        # Always include the "type" field if present,
-        # even if it was not set (important for style merging).
-        if "type" in self.__dict__:
-            d["type"] = self.__dict__["type"]
-
-        return d
-
     @cached_property
     def _cached_hash(self) -> int:
         """This is safe because all style fragments are immutable."""
         return hash(self)
+
+    def mergeable_dump(self) -> dict[str, object]:
+        diff: dict[str, object] = {}
+
+        for field_name, value in self:
+            field = self.model_fields[field_name]
+            field_default = field.default
+
+            if field_name not in self.model_fields_set:
+                continue
+
+            if isinstance(value, StyleFragment):
+                if isinstance(field_default, StyleFragment):
+                    # Value and default are both fragments
+                    sub_diff = value.mergeable_dump()
+                    if isinstance(field.discriminator, str):
+                        sub_diff[field.discriminator] = getattr(value, field.discriminator)
+                    if sub_diff:
+                        diff[field_name] = sub_diff
+                else:
+                    # Value is a fragment, default is a primitive
+                    diff[field_name] = value.mergeable_dump()
+            else:
+                # Not a fragment, so it's a primitive
+                diff[field_name] = value
+
+        return diff
 
 
 class Color(NamedTuple):
@@ -390,6 +404,9 @@ class BorderKind(Enum):
         right_bottom="*",
     )
 
+    def __repr__(self) -> str:
+        return f"BorderKind.{self.name}"
+
 
 class JoinedBorderParts(NamedTuple):
     vertical: str
@@ -543,6 +560,9 @@ class JoinedBorderKind(Enum):
         horizontal_vertical="╬",
     )
 
+    def __repr__(self) -> str:
+        return f"JoinedBorderKind.{self.name}"
+
 
 class BorderEdge(Enum):
     Top = "top"
@@ -553,7 +573,7 @@ class BorderEdge(Enum):
 
 class Border(StyleFragment):
     kind: BorderKind = Field(default=BorderKind.Light)
-    style: CellStyle = Field(default_factory=CellStyle)
+    style: CellStyle = Field(default=CellStyle())
     edges: frozenset[BorderEdge] = frozenset({BorderEdge.Top, BorderEdge.Bottom, BorderEdge.Left, BorderEdge.Right})
     contract: int = Field(default=0)
 
@@ -585,7 +605,35 @@ class Typography(StyleFragment):
     wrap: Literal["none", "paragraphs"] = "none"
 
 
+class Relative(StyleFragment):
+    """
+    Relative positioning is relative to the parent element's content box.
+    Elements occupy space and are laid out next to their siblings according
+    to the parent's layout direction.
+    """
+
+    type: Literal["relative"] = "relative"
+    x: int = 0
+    y: int = 0
+
+
 class Absolute(StyleFragment):
+    """
+    Absolute positioning is relative to the parent element's content box,
+    but the element does not occupy space.
+    """
+
+    type: Literal["absolute"] = "absolute"
+    x: int = 0
+    y: int = 0
+
+
+class Fixed(StyleFragment):
+    """
+    Fixed positioning is relative to the screen's top-left corner `(0, 0)`.
+    """
+
+    type: Literal["fixed"] = "fixed"
     x: int = 0
     y: int = 0
 
@@ -593,7 +641,7 @@ class Absolute(StyleFragment):
 class Flex(StyleFragment):
     type: Literal["flex"] = "flex"
     direction: Literal["row", "column"] = "row"
-    position: Literal["relative"] | Absolute = "relative"
+    position: Relative | Absolute | Fixed = Field(default=Relative(), discriminator="type")
     weight: PositiveInt | None = 1
     align_self: Literal["none", "start", "center", "end", "stretch"] = "none"
     justify_children: Literal[