+++ counterweight/styles/utilities.py
@@ -1,13 +1,16 @@
-from counterweight.styles import Margin, Style
-from counterweight.styles.styles import (
+from counterweight.styles import (
     Absolute,
     Border,
     BorderEdge,
     BorderKind,
     CellStyle,
     Color,
+    Fixed,
     Flex,
+    Margin,
     Padding,
+    Relative,
+    Style,
     Typography,
 )
 
@@ -1958,5 +1961,13 @@ text_justify_right = Style(typography=Typography(justify="right"))
 # Stop generated
 
 
+def relative(x: int = 0, y: int = 0) -> Style:
+    return Style(layout=Flex(position=Relative(x=x, y=y)))
+
+
 def absolute(x: int = 0, y: int = 0) -> Style:
     return Style(layout=Flex(position=Absolute(x=x, y=y)))
+
+
+def fixed(x: int = 0, y: int = 0) -> Style:
+    return Style(layout=Flex(position=Fixed(x=x, y=y)))